[FEAT] orm: add chain-linkage between models
Now, with this commit, it will be possible to
have linkages by models. For example, if you define

class A:
  _table = 't_a'

class B:
  _table = 't_b'

  a_id = Fields.reference('t_a')

before, to get the class A from the B, we had to do
smth like that:

1. b = B(9)
2. env['t_a'](b.a_id)

now, we can just do:

b.a_id

and it will give you access to the entire A object.

For assign a new value, you can just pass the new saved
model or just the ID.