[FEAT] wallet: new purchase
Now, with this commit, the purchase method is
working using the window buy, set in the wallet
in minutes. If the current price is less than the
rebased price AND the last purchase was before a
window specified into the wallet, then a purchase
can be done. This commit introduces some methods
in the wallet, like the:

* cash_amount(): The BRL balance available for
buy new btc

* invested_amount(): The total BRL invested,
considering the money that we have received.

* can_buy(): analyser considering the window
time shift

* cash_holding(): a directed link to the main
holding - in BRL.